+++ 03-blue-green/deployment.yaml.template
@@ -11,15 +11,23 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-apiVersion: v1
-kind: Service
+
+apiVersion: extensions/v1beta1
+kind: Deployment
 metadata:
-  name: hello-node
-spec:
-  ports:
-    - port: 80
-      targetPort: 8080
-      protocol: TCP
-  selector:
+  name: hello-node-blue
+  labels:
     name: hello-node-blue
-  type: LoadBalancer
+spec:
+  replicas: 5
+  template:
+    metadata:
+      labels:
+        name: hello-node-blue
+    spec:
+      containers:
+      - name: hello-node
+        image: gcr.io/<PROJECT_ID>/imageflipper-app:<VERSION>
+        imagePullPolicy: Always
+        ports:
+        - containerPort: 8080
