+++ 03-blue-green/service.yaml
@@ -11,22 +11,15 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-apiVersion: extensions/v1beta1
-kind: Deployment
+apiVersion: v1
+kind: Service
 metadata:
-  name: hello-node-green
-  labels:
-    name: hello-node-green
+  name: hello-node-test
 spec:
-  replicas: 5
-  template:
-    metadata:
-      labels:
-        name: hello-node-green
-    spec:
-      containers:
-      - name: hello-node
-        image: gcr.io/<PROJECT_ID>/imageflipper:1.0
-        imagePullPolicy: Always
-        ports:
-        - containerPort: 8080
+  ports:
+    - port: 80
+      targetPort: 8080
+      protocol: TCP
+  selector:
+    name: hello-node-blue
+  type: LoadBalancer
