+++ CONTRIBUTORS
@@ -1,1 +1,2 @@
 Sandeep Dinesh (@thesandlord)
+Robert Kubis (@hostirosti)
