--- blue-green/deployment.yaml
@@ -1,32 +1,0 @@
-# Copyright 2017 Google Inc. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-# 
-#     http://www.apache.org/licenses/LICENSE-2.0
-# 
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-apiVersion: extensions/v1beta1
-kind: Deployment
-metadata:
-  name: hello-node-blue
-  labels:
-    name: hello-node-blue
-spec:
-  replicas: 5
-  template:
-    metadata:
-      labels:
-        name: hello-node-blue
-    spec:
-      containers:
-      - name: hello-node
-        image: gcr.io/<PROJECT_ID>/imageflipper:2.0
-        imagePullPolicy: Always
-        ports:
-        - containerPort: 8080
