--- blue-green/index.html
@@ -1,26 +1,0 @@
-<!--
-  Copyright 2017, Google, Inc.
- Licensed under the Apache License, Version 2.0 (the "License");
- you may not use this file except in compliance with the License.
- You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
- Unless required by applicable law or agreed to in writing, software
- distributed under the License is distributed on an "AS IS" BASIS,
- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- See the License for the specific language governing permissions and
- limitations under the License.
--->
-<!DOCTYPE html>
-<html>
-<head>
-  <title>Flipper</title>
-</head>
-<body>
-  <form id="uploadForm" enctype="multipart/form-data" action="/api/photo" method="post">
-    <input type="file" name="userPhoto" />
-    <input type="submit" value="Upload Image" name="submit" />
-  </form>
-</body>
-</html>
