--- blue-green/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "imageflip",
-  "version": "2.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node index.js"
-  },
-  "dependencies": {
-    "express": "4.13.3",
-    "multer": "1.1.0",
-    "request": "^2.81.0"
-  },
-  "keywords": [],
-  "author": "Google Inc",
-  "license": "Apache-2.0"
-}
