+++ imageflipper-app/LICENSE
