+++ imageflipper-app/index.html
@@ -1,0 +1,45 @@
+<!--
+  Copyright 2017, Google, Inc.
+ Licensed under the Apache License, Version 2.0 (the "License");
+ you may not use this file except in compliance with the License.
+ You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+ Unless required by applicable law or agreed to in writing, software
+ distributed under the License is distributed on an "AS IS" BASIS,
+ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ See the License for the specific language governing permissions and
+ limitations under the License.
+-->
+<!DOCTYPE html>
+<html>
+
+<head>
+    <title>Flipper</title>
+</head>
+
+<body onload="version()">
+    <form id="uploadForm" enctype="multipart/form-data" action="/api/photo" method="post">
+        <input type="file" name="userPhoto" />
+        <input type="submit" value="Upload Image" name="submit" />
+    </form>
+    </br>
+
+    Proudly served by version <span id="version"></span> :)
+    <script type="text/javascript" language="javascript">
+        function version() {
+            fetch('./api/version')
+                .then(function(response) {
+                    response.text().then(function(data) {
+                        document.querySelector('#version').innerHTML = data;
+                    })
+                })
+                .catch(function(err) {
+                    document.querySelector('#version').innerHTML = 'N/A';
+                });
+        }
+    </script>
+</body>
+
+</html>
