+++ imageflipper-app/package.json
@@ -1,0 +1,19 @@
+{
+    "name": "imageflip",
+    "version": "1.0.0",
+    "description": "",
+    "main": "index.js",
+    "scripts": {
+        "test": "echo \"Error: no test specified\" && exit 1",
+        "start": "node index.js"
+    },
+    "dependencies": {
+        "express": "4.13.3",
+        "multer": "1.1.0",
+        "sharp": "^0.18.4",
+        "request": "^2.83.0"
+    },
+    "keywords": [],
+    "author": "Google Inc",
+    "license": "Apache-2.0"
+}
