+++ imageflipper-service/service.yaml
@@ -14,11 +14,11 @@
 apiVersion: v1
 kind: Service
 metadata:
-  name: annotate
+  name: flipper-service
 spec:
   ports:
     - port: 80
       targetPort: 8080
       protocol: TCP
   selector:
-    name: annotate
+    name: flipper-service
